Add ssz.HashRoot interface composition to BeaconBlock interface, move fork specific logic into it's own files
