+++ proto/prysm/v1alpha1/block/block_interfaces.go
@@ -38,6 +38,7 @@ type BeaconBlock interface {
 	Proto() proto.Message
 	ssz.Marshaler
 	ssz.Unmarshaler
+	ssz.HashRoot
 	Version() int
 }
