+++ proto/prysm/v1alpha1/wrapper/BUILD.bazel
@@ -4,6 +4,9 @@ go_library(
     name = "go_default_library",
     srcs = [
         "beacon_block.go",
+        "beacon_block_altair.go",
+        "beacon_block_bellatrix.go",
+        "beacon_block_phase0.go",
         "metadata.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/proto/prysm/v1alpha1/wrapper",
@@ -14,6 +17,7 @@ go_library(
         "//proto/prysm/v1alpha1/block:go_default_library",
         "//proto/prysm/v1alpha1/metadata:go_default_library",
         "//runtime/version:go_default_library",
+        "@com_github_ferranbt_fastssz//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_eth2_types//:go_default_library",
         "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
@@ -23,7 +27,12 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["beacon_block_test.go"],
+    srcs = [
+        "beacon_block_altair_test.go",
+        "beacon_block_bellatrix_test.go",
+        "beacon_block_phase0_test.go",
+        "beacon_block_test.go",
+    ],
     deps = [
         ":go_default_library",
         "//encoding/bytesutil:go_default_library",
